refactor : 계산방식 변경 및 딜러 테스트를 유저테스트로 이동
