+++ src/main/kotlin/blackjack/domain/player/Player.kt
@@ -17,18 +17,11 @@ abstract class Player(val name: String, initCards: List<Card>) {
     val cards: Cards
         get() = _cards
 
-    var income = Money()
-        private set
-
-    fun addIncome(money: Money) {
-        income += money
-    }
-
     init {
         require(name.isNotEmpty()) { ErrorMessages.NAME_IS_EMPTY }
     }
 
-    fun hit(card: Card) {
+    private fun hit(card: Card) {
         _cards.addCard(card)
     }
